Fix custom search problem, changed exclude behaviour to use OR instead of AND for excludes
Since the upgrade the custom search didn't exclude properly, for example:

	len(Image.objects.filter(Q(tagging_status__slug__exact='obs'), ~Q(tagging_status__slug__exact='coords') | ~Q(tagging_status__slug='no_coords')))

returns 1563, but:

	len(Image.objects.filter(tagging_status__slug='obs').exclude(Q(tagging_status__slug='coords') | Q(tagging_status__slug='no_coords')))

return 0, which is the expected result. The code was updated accordingly.

Moreover, the code was using AND between exclude queries, and OR between filter
queries, but the interface was showing OR. The code was updated to use OR between
both filter and exclude queries